support xpu op in separate file  * test=kunlun
